Rename scenes feature into shape-from-mesh
